:bug: フックのエラーを解消
