+++ src/components/clubPage/edit.tsx
@@ -3,8 +3,16 @@ import { forbidden, notFound, unauthorized } from "next/navigation";
 import { use } from "react";
 import ClubEdit from "./editComponent";
 
-export default function EditClub({ id }: { id: string }) {
-  const club = use(getClubById(id));
+export default function EditClub({
+  id,
+  apiBase,
+  sessionID,
+}: {
+  id: string;
+  apiBase: string;
+  sessionID: string | undefined;
+}) {
+  const club = use(getClubById(id, apiBase, sessionID));
   if (typeof club == "string") {
     switch (club) {
       case "forbidden":
